Fixed CTA bg colours test
